Smart-Scheduler/backend/requirements.txt
Flask==3.0.3
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.1
Flask-Limiter==3.7.0
APScheduler==3.10.4
pymongo==4.7.2
bcrypt==4.1.3
marshmallow==3.21.3
firebase-admin==6.5.0
gunicorn==22.0.0
python-dotenv==1.0.1
